+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Change categories:
 - dsp: platform: Add `possible_states` iterator for the State Sensor PDR
   - `foreach_pldm_platform_state_sensor_pdr_possible_states()`
   - `foreach_pldm_platform_state_sensor_pdr_states()`
+- dsp: platform: Add `encode_pldm_platform_set_state_sensor_enables_req()`
 - libpldm++ support for fw update pkg v1.1.0
 - libpldm++ support for fw update pkg v1.2.0
 - libpldm++ support for fw update pkg v1.3.0